Fix compile warnings

+++ src/signals.h
@@ -105,7 +105,7 @@ namespace Beeps
 
 			uint nframes () const
 			{
-				return nFrames_;
+				return (uint) nFrames_;
 			}
 
 			const Float* data () const
